// src/entities/profile/model/type/profile.ts
import { Country } from 'entities/country'
import { Currency } from 'entities/currency'

export interface Profile {
  firstname?: string
  lastname?: string
  age?: number
  currency?: Currency
  country?: Country
  city?: string
  username?: string
  avatar?: string
}

export interface ProfileSchema {
  data?: Profile
  form?: Profile
  isLoading: boolean
  error?: string
  readonly: boolean
}